NI-4922 22.1.10 Alchemy Port Clean Up
This commit updates the iso3166-2 database for the alchemy pycountry
port to be up to date. Previous updates have cherry picked changes but
this has caused some updates that took place between deprecation of
Python 2/3 support and Administrate forking PyCountry to be missed,
causing some province codes to be out of date while others have stayed
up to date.